+++ src/tilesource.c
@@ -30,8 +30,8 @@
 /*
 #define DEBUG_VERBOSE
 #define DEBUG_MAKE
- */
 #define DEBUG
+ */
 
 #include "nip4.h"