+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - better power of two detection in openslideload
+- fix tile edges with gtk_snapshot_set_snap() [kleisauke]
 
 ## 9.0.11 2025/07/21